requirements.txt
astroid==2.5.1
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
caffeine==2.9.4
certifi==2020.4.5.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==3.0
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro==1.5.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
ewmh==0.1.6
fasteners==0.14.1
Flask==1.1.2
future==0.18.2
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.0
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==21.3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lazy-object-proxy==1.5.2
lockfile==0.12.2
louis==3.14.0
macaroonbakery==1.3.1
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.0.3
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pdf2image==1.14.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.6
protobuf==3.12.3
pycairo==1.16.2
pycups==2.0.1
Pygments==2.3.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
PySimpleGUI==4.37.0
pytesseract==0.3.7
python-apt==2.1.3+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.37
python-distutils-extra==2.39
python-Levenshtein==0.12.0
python-xlib==0.27
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
reportlab==3.5.47
requests==2.23.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==3.1.2
Send2Trash==1.5.0
simplejson==3.17.0
six==1.15.0
ssh-import-id==5.10
systemd-python==234
terminado==0.8.2
testpath==0.4.4
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
ulauncher==5.10.0
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
wadllib==1.3.4
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==1.0.1
widgetsnbextension==2.0.0
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0